refactor: extract table utilities
